Re-organising surface interaction into shape - will remove surface interactable trait later

+++ colour/src/samplers/sobol.rs
@@ -134,7 +134,7 @@ fn sobol_sample_f64(mut index: u64, dimension: usize, scramble: u64) -> f64 {
         "Too many sobol dimensions consumed"
     );
 
-    let mut result = scramble & (!-((1 << data::sobol::MATRIX_SIZE) as i64)) as u64;
+    let mut result = scramble & (!-(1i64 << data::sobol::MATRIX_SIZE)) as u64;
 
     for i in dimension * data::sobol::MATRIX_SIZE.. {
         if index & 1 != 0 {
@@ -146,5 +146,5 @@ fn sobol_sample_f64(mut index: u64, dimension: usize, scramble: u64) -> f64 {
         }
     }
 
-    ((result as f64) * (1.0 / ((1 << data::sobol::MATRIX_SIZE) as f64))).min(f64::ONE_MINUS_EPSILON)
+    ((result as f64) * (1.0 / ((1i64 << data::sobol::MATRIX_SIZE) as f64))).min(f64::ONE_MINUS_EPSILON)
 }
